requirements.txt
appdirs==1.4.4
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.1
Django==1.11.29
djangorestframework==3.9.4
filelock==3.0.12
gunicorn==19.10.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
pathlib2==2.3.5
Pillow==6.2.2
pytz==2020.1
scandir==1.10.0
singledispatch==3.4.0.3
six==1.15.0
typing==3.7.4.3
virtualenv==20.0.31
zipp==1.2.0
